Add amounts to bars on mouse hover
The bars is the graph now show the amount when the user hovers over them. If the bar is longer than the text, the amount will be drawn in white on the left part of the bar. Otherwise, the it will be drawn in brown to the right of the bar.